fix(parquet store gateway): obey query sharding matchers (#12018)
Inefficient, but at least correct query sharding. The new test on
sharding fails on the base branch.

It's not trivial to add caching to the hashes like the main path does,
because we don't have a `SeriesRef` to use as a cache key at the block
level (to match what the main path does). We could in theory use
something like the row number in the parquet file, but we don't have
easy access to that in this part of the code. In any case, the priority
right now is correctness, we'll work on optimizing later as appropriate.

For referece, see how query sharding is handled on the main path:
https://github.com/grafana/mimir/blob/604775d447c0a9e893fa6930ef8f2d403ebe6757/pkg/storegateway/series_refs.go#L1021-L1047