operations: Add knob for controlling ingest-storage producer record version (#11969)
#### What this PR does

Exposes a config knob that controls and validates the experimental
`ingest-storage.kafka.producer-record-version` arg. We default to 0,
which is the current default if unsupplied.

It seems we don't support any `-ingest-storage.*` controls in the helm
chart, so I didn't touch that.

#### Which issue(s) this PR fixes or relates to

Contrib https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.