Update IngesterHasNotShippedBlocks to be block-builder aware (#12024)
#### What this PR does

In #11872 we added the
`cortex_blockbuilder_tsdb_last_successful_compact_and_upload_timestamp_seconds`
metric, that the block-builder exposes to track its shipment of blocks.

In this PR we update the ingester alerts to not trigger if the mentioned
metric exists. This is so we could safely stop the ingesters from
uploading the blocks to the object store, while the block-builder is
doing that.

Also, the PR updates related runbooks. It removes the references to
alerts and metrics of the legacy stand-alone mode, and adds the new one.

---------

Signed-off-by: Vladimir Varankin <vladimir.varankin@grafana.com>