rw2: Validate that the first symbol of the RW2 Symbols field, and common labels, is an empty string (#11953)
#### What this PR does

Per the [Remote-Write 2.0
spec](https://prometheus.io/docs/specs/prw/remote_write_spec_2_0/#symbols):

> The first element of the symbols table MUST be an empty string, which
is used to represent empty or unspecified values such as when
Metadata.unit_ref or Metadata.help_ref are not provided.

Obviously Prometheus [follows its own
spec](https://github.com/prometheus/prometheus/blob/74aca682b77d66f5761bd68459f3b570ec34ce2d/prompb/io/prometheus/write/v2/symbols.go#L27-L29),
but other clients might send a request that's out of spec. If a client
does this, we accept their request, but we might store garbage.
Typically, this looks like mangling their metric metadata into a random
label name used elsewhere in the request when a field is unsupplied,
possibly for an unrelated series.

We cannot decipher the meaning of such a request. Say a metadata unit
has a reference value of 0, we can't tell whether the client intended to
set it to the first symbol or to leave it blank. We don't have enough
info to try to interpret it. So, this PR proposes rejecting such
requests completely, since per the spec they are not valid anyway.

The ingest-storage record format V2 had a worse version of this bug,
which happens even for valid requests! In that format, we reserve the
first N symbols for common symbols. An unsupplied metadata field
serializes as the value `0`, which is in the common symbols space. We
add a mandate that common symbols must start with empty string just like
regular symbols so we still behave as expected when fields are empty.

**Alternatives Considered**

We could instead try to partially reject metadata, when we detect this
problem. Empty strings are rare, and only frequently used in metadata -
if a request has a bad symbols table, we could only reject TimeSeries
containing a 0 reference/empty string somewhere in it, and still
partially accept all other timeseries that are unaffected or have no
metadata. This is more complex, and didn't feel worth it given the
language in the spec, but better follows [Postel's
Law](https://en.wikipedia.org/wiki/Robustness_principle). We can go that
route if reviewers prefer.

#### Which issue(s) this PR fixes or relates to

Contrib: https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>
Co-authored-by: Marco Pracucci <marco@pracucci.com>