Fix issues with `mimirtool remote-read export` and long time ranges (#12025)
#### What this PR does

This PR fixes a number of issues in `mimirtool remote-read export` that
could occur when querying a range longer than an hour, or that crossed
block boundaries (which were hardcoded to 2h in size). In particular,
queries that crossed block boundaries could omit some output blocks.

Key changes:

- Output blocks will now be 2h in length, not 1h, and this can be
configured with `--block-size`.
- Time ranges that cross block boundaries are now handled correctly,
including for cases where some series appear in only some output blocks.
- Long time ranges are now broken up into intervals that match the
output block size, reducing the amount of data that must be held in
memory by `mimirtool` and simplifying the logic of producing blocks in
this case.
- An alternative solution would be to create multiple block writers, one
for each output block, and append to the correct one based on the
samples received, but this would require all received samples to be held
in memory until the last one is received.
- Details of blocks are now reliably printed to the console, rather than
sometimes being incomplete or omitted entirely, or failing with
`read/write on closed pipe`.

#### Which issue(s) this PR fixes or relates to

Related to #11404

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>
(cherry picked from commit c5137e360f8bfad83a2f80d18b3579fbb8d8af24)