Reduce TestConcurrentFetchers flakyness likelihood (#12028)
#### What this PR does

In this PR I'm (hopefully) reducing the likehood of
`TestConcurrentFetchers` flakyness, in particular focusing on `respect
maximum buffered bytes limit` case, which is the most flaky one. I'm not
calling it a fix because, as usual, I haven't been able to reproduce it
locally, but the changes have been originated from the behaviour I've
observed in CI execution logs. Here you can see the [logs of a failed
execution](https://github.com/grafana/mimir/issues/11859#issuecomment-3052493066).

In this PR:

- Do not run the `respect maximum buffered bytes limit` concurrently.
I've seen in logs that there's essentially nothing wrong, but it takes
longer than 2s to reach stabilization (and from logs I can see it was
still fetching records).
- Increase `waitForStableBufferedRecords()` timeout from 2s to 5s.
- Improve `waitForStableBufferedRecords()` for the case when the
buffered records is 0 when it gets called. Previously there was no
comparison, now it waits 1 tick (100ms) and then run the comparison.
- Produce records in batches, like in a real world scenario. Previously
in the test records were produced in individual batches, which means
kfake was actually returning 1 record at a time, regardless how many we
requested (behaviour is clearly visible in logs). This also makes the
consumption (fetcher) much faster because it fetches 1 batch at a time,
instead of 1 record at a time from kfake.
- To produce records, use the real kafka client config we use in
distributors, but without lingering (to have tests run faster).

Time will tell if this helps.

#### Which issue(s) this PR fixes or relates to

Part of https://github.com/grafana/mimir/issues/11859

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Marco Pracucci <marco@pracucci.com>