write path: handle created timestamp (#11977)
#### What this PR does

Modify the ingester and block builder to accept and pass the Created
timestamp field from incoming write requests to TSDB.

Note1: if samples aren't ordered by time in a remote write requests,
then the result is undefined.

Note2: as with regular samples, Mimir will have non-deterministic
behavior if there's a sample and a created timestamp with the same
timestamp ingested on two different ingesters for whatever reason.
During query and block building the resolution of such conflicting
samples has always been non-deterministic, as there is no way to know at
that point which sample is the correct one.

#### Which issue(s) this PR fixes or relates to

Fixes #11976

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>
Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>