Optimise label names API by removing __name__!="" matcher when redundant (#12054)
#### What this PR does

This PR introduces a new labels query optimizer middleware for the
query-frontend that automatically removes
the ` __name__!=""` matcher from label names queries when the matcher is
redundant. The PR also includes a bug
fix for the limit parameter validation in label queries: the value 0 was
not supported previously, but it's actually a valid parameter in
Prometheus (means unlimited).

I've tested the following matchers, comparing the result with and
without the optimizer enabled and I've got the same exact results:
- No matchers
- With `{}` matcher (invalid)
- With `{__name__!=""}` matcher
- With `{__name__!="", container="mimir-write"}` matcher
- With `{__name__!="", __name__="up"}` matcher
- With `{__name__!=""}` and `{__name__=~"a.*"}` matchers
- With `{__name__!="", __name__=~"a.*"}` matcher

This PR only address the label names API. I will then open a PR for the
label values.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>