Fix possible panic when handling OTLP push errors
While developing a different feature I encountered a panic when a gRPC status error bubbled up through the middlewares to the OTLP handler, which tried to write the gRPC status code as an HTTP status code.

I can't see how to reproduce this without that feature, but in any case the handler should be prepared to handle a gRPC error from the complex stack of distributor's middlewares.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>