Fix possible panic when handling OTLP push errors (#12072)
#### What this PR does

While developing a different feature I encountered a panic when a gRPC
status error bubbled up through the middlewares to the OTLP handler,
which tried to write the gRPC status code as an HTTP status code.

I can't see how to reproduce this without that feature, but in any case
the handler should be prepared to handle a gRPC error from the complex
stack of distributor's middlewares.

#### Which issue(s) this PR fixes or relates to

Fixes: https://github.com/grafana/mimir/issues/12019

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>