Update dskit to skip tracing configuration when no env is defined (#12074)
#### What this PR does

This updates dskit to include https://github.com/grafana/dskit/pull/723

This will stop configuring the tracing libraries when no tracing
environment variables were provided, as by default OTel libraries will
try to export spans to `localhost` and complain because it would reject
them.

#### Which issue(s) this PR fixes or relates to

Related to the migration to OTel tracing.

#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>