feature(parquet): Implement store-gateway limits (#12040)
This PR is based on the upstream work
https://github.com/prometheus-community/parquet-common/pull/81

The idea is to implement a set of basic quota limiters that can protect
us against potential bad queries for the gateways.

Note we had to bring bits of the code available in the querier in
upstream because we have our own chunk querier in Mimir.

---------

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>