Query-frontend: fix auth propagation in labels query optimizer (#12076)
#### What this PR does

In https://github.com/grafana/mimir/pull/12054 and
https://github.com/grafana/mimir/pull/12066 I've done a mistake that I
haven't noticed when testing it in the local dev env: the
`X-Scope-OrgID` was not propagated when a query is optimised. In this PR
I'm fixing it by using `user.InjectOrgIDIntoHTTPRequest()`: this follows
the same strategy used in other places in the query-frontend (i.e.
instant/range queries, and active series sharding). In addition, I'm
also adding an integration test, that tests idempotency of the optimizer
comparing label names and values query responses with and without
optimizer enabled.

https://github.com/grafana/mimir/pull/12054 was merged last week: should
we back port the fix to the last week release? No, because the optimizer
is experimental and disabled by default. When disabled, no request is
rewritten, and so the auth bug doesn't show up.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>