Improve docs for blocked requests, blocked queries and limited queries config options (#12053)
#### What this PR does

This PR improves how the configuration of blocked requests, blocked
queries and limited queries is displayed in the docs.

In particular:

* all fields are now shown as they would be for any other config block,
rather than relying on a comment
* an example has been added for how to configure a blocked request

To achieve this, I've made two changes to the config doc generator:

* maps of structs are now shown expanded, instead of as `map of string
to package.SomeStruct`
* examples (if available) are now shown for all types of fields

I've also changed the implementation of `ExampleDoc` for each type to
use the Go type rather than a generic map, so these should stay in-sync
with the actual configuration fields. I changed the blocked requests,
blocked queries and limited queries slices to not use pointers, as the
pointers seemed unnecessary and keeping them would have required further
changes to the config doc generator.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [x] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.