chore: Update dskit (#12099)
#### What this PR does

This updates dskit to bring the latest features needed for the
usage-tracker branch. Updating main so I have to deal with less
conflicts in the branch.

This has no user-facing changes.

#### Which issue(s) this PR fixes or relates to

None

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Oleg Zaytsev <mail@olegzaytsev.com>