Query-frontend: optimise labels API by removing =~'.*' matcher when redundant (#12080)
#### What this PR does

This is a follow up of https://github.com/grafana/mimir/pull/12054 and
https://github.com/grafana/mimir/pull/12066, implementing the suggestion
received by @bboreham
[here](https://github.com/grafana/mimir/pull/12054#discussion_r2200625105).
The idea is to filter out `label=~".*"` matcher when it's redundant: in
Prometheus 3 (and Mimir now) `.*` matches any string, including new
line, so the `.*` matcher literally matches every series.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>
Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>