Refactoring: move X-Scope-OrgID HTTP header injection into EncodeLabelsSeriesQueryRequest() and EncodeMetricsQueryRequest() (#12100)
#### What this PR does

This PR addresses the comment received
[here](https://github.com/grafana/mimir/pull/12076#discussion_r2205602043).
The idea is to move the X-Scope-OrgID HTTP header injection into
`EncodeLabelsSeriesQueryRequest()` and `EncodeMetricsQueryRequest()`
instead of doing it in the caller, right after encoding requests. The
`EncodeLabelsSeriesQueryRequest()` and `EncodeMetricsQueryRequest()` are
only called in places where `user.InjectOrgIDIntoHTTPRequest()` is
called right after (and same in GEM, that I will take care of updating
if this PR is accepted).

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Marco Pracucci <marco@pracucci.com>