Refactoring: remove querymiddleware.Codec interface (#12101)
#### What this PR does

`querymiddleware.Codec` is an interface, but there's no need for an
interface nowadays. We introduced the `Codec` interface years ago in
Cortex because, at that time, Loki was vendoring query-frontend from
Cortex and they needed a different implementation. Nowadays the only
codec is `prometheusCodec` (I also checked GEM).

In this PR I propose to remove the `Codec` interface and directly expose
the codec implemention instead. This removes an unnecessary indirection
and makes navigating the code easier.

If this PR is accepted, I will take care of replacing the `Codec` usage
in GEM.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.