Enforce consistent formatting of Protobuf files (#12036)
This PR adds Buf to our build image, and uses it to enforce consistent
formatting of Protobuf files.

While the generated Go files have changed, all of these changes are in
the file descriptor, and have no impact on Mimir's behaviour.

(none)

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.