fix(frontend): queries with duration expression (#12038)
#### What this PR does

The sharding, splitting middlewares do not handle duration expressions
in PromQL queries. Any offset, range, step that would be calculated from
a duration expression like (e.g. `metric offset (2+2)`) is treated as if
the expression wasn't there and the value of the offset, range, step was
0/default (e.g. `metric offset 0`).

Currently all such expressions can be pre-calculated based on the query
step. Add a middleware to pre-calculate the offsets, ranges, steps.

#### Which issue(s) this PR fixes or relates to

Fixes N/A

#### Checklist

- [x] Tests updated.
- [N/A] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [N/A]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>
Co-authored-by: Jesus Vazquez <jesusvazquez@users.noreply.github.com>
Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>