Allow cost attribution label config to include output labels (#12035)
#### What this PR does

This PR enables configuration of an output label for each defined cost
attribution label. A new `cost_attribution_labels_structured` YAML
element is added to the `Limits` struct with the structure:

```go
type Label struct {
	Input, Output string
}
```

This allows both an "input" label that refers to metrics actually in use
in the system, and an "output" label which is what will be emitted at
scrape time in the usage metrics.

`cost_attribution_labels_structured` is preferred over
`cost_attribution_labels` if both are present.