Standardise non-regex patterns in query blocking upon loading of config (#12102)
#### What this PR does

Make query blocking easier to use by not needing users to run `mimirtool
promql format` themselves, instead applying a similar transformation to
the patterns upon loading.

This is skipped for regex patterns as they may not compile into proper
PromQL expressions.

Should be fairly efficient as it is only parsed upon loading and not
every time it is compared.

Currently only applies to query blocking, not sure if we also want this
for query limiting or others.

To keep things working similarly to before, we don't return any error if
the PromQL query is invalid, and we just keep the string for comparison,
but it will never match anything. If we would rather change this, we
could move it to the validation step and return an error if a pattern is
invalid. Or we could silently remove the useless string from being
checked without returning an error.

Tested to work locally.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [x] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>