Alertmanager: Support kind in Templates coming from Grafana (#12089)
#### What this PR does
Updates GrafanaAlertmanagerConfig with a new field `templates` of type
`[]definition.PostableApiTemplate`. This field will eventually replace
map `template_files` but for the sake of backward compatibility it is
still supported. Both fields are merged into single collection, which is
then propagated downstream.

This will let us control which fields are populated on Grafana side, and
transition to the new field gracefully.


This PR is built on top of the refactoring
https://github.com/grafana/mimir/pull/12052

No Mimir Alertmanager related  APIs or models are affected.