jsonnet, helm: Reduce distributor GOMAXPROCS (#12150)
#### What this PR does

Reduce the calculated value of `GOMAXPROCS` for distributors to be
closer to the requested CPU. This upstreams a change we have been using
and tesing internally.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [X] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>