Block-builder-scheduler: fix bugs in handling of partitions with no commit (#12130)
There was a bug introduced in #11867 where if a block-builder-scheduler
encounters a partition in this state:
* the partition has an end offset > 0
* the block-builder consumer group has **no commit** for that partition

(This can happen if the partition hasn't been written to since before
the scheduler or its consumer group existed in that environment.)

then:
* the block-builder-scheduler would store a committed offset of `0` in
the partition state

which leads to:
1. the block-builder-scheduler would erroneously detect an offset gap
the next time a job is created for that partition. (because only a job
spec with `startOffset=0` would be allowed by the gap checker.)
2. the scheduler would emit a value of `0` for the partition in the
`cortex_blockbuilder_scheduler_partition_committed_offset` gauge.

This happened because commits _were_ stored in a partition-keyed map
that was only populated if either the consumer group had a commit for
the partition, or if block-builder completed a job on the partition. But
now, they're stored in the `partitionState` struct with a 0-defaulting
commit offset.

To fix this, I am introducing a special state of `offsetEmpty` (`-1`) to
the offsets that are handled specially: The gap detector will allow a
transition from `offsetEmpty` to anything else as a one-time exception.

This PR also adds a new gauge
`cortex_blockbuilder_scheduler_partition_planned_offset` for observing
the latest planned offset per partition.

(cherry picked from commit 0a75686b7a7b555ee8e9bc15458b1899e2b067b5)