Alertmanager to store fingerprint of configuration instead of structure (#12087)
#### What this PR does
This PR improves multitenant Alertmanager to store a fingerprint of user
configuration instead of a full structure. The structure uses this field
to compare current configuration with one loaded from the storage and
skip applying the configuration if it's not changed.
I decided to implement this function on my own and add necessary tests
to assert the correctness of the fingerprint and protect for unintended
modifications in future.