ingester: Stagger head compaction intervals when running ingesters in multiple zones (#12090)
#### What this PR does

Attempts to calculate staggered intervals for ingesters in each zone to
avoid ingesters across all zones to do head compaction at the same time,
which in turn would make strong consistency queries fail.

#### Which issue(s) this PR fixes or relates to

Fixes #12081

TODO:

 - [x] integration test
 - [x] fix unit test
- [ ] determine whenever we should take into account if compactions go
over the interval duration and we have a risk of aligning compactions
again

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [x]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: gotjosh <josue.abreu@gmail.com>

---------

Signed-off-by: gotjosh <josue.abreu@gmail.com>