refactor: remove ability to set downstream-url in query-frontend (#12191)
#### What this PR does

Remove the ability to set a downstream-url in the query-frontend which
allowed the query-frontend to act as a frontend for any Prometheus
compatible server. The functionality is not used internally and
complicates maintenance of the query-frontend.

#### Which issue(s) this PR fixes or relates to

Fixes #11883

#### Checklist

- [X] Tests updated.
- [X] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>