Remove use of `thanos-io/objstore` fork no longer needed (#12211)
#### What this PR does

The use of a fork of `thanos-io/objstore` was originally introduced in
https://github.com/grafana/mimir/pull/11546 to work around the bug fixed
in https://github.com/thanos-io/objstore/pull/181. This fix has been
merged as part of https://github.com/thanos-io/objstore/pull/196, so the
fork is no longer necessary.

The latest version of `thanos-io/objstore` depends on updated versions
of some of its dependencies, so these are being bumped as part of this
PR as well.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.