operations: Allow ingest-storage record version to be increased to v2 (#12190)
#### What this PR does

Allows V2 as a producer record version as of
https://github.com/grafana/mimir/pull/12060.
V2 is still considered unstable at the time of writing. We might make
breaking changes to the format.

Thus, we include a guard where you have to explicitly opt-in to
experimental formats before setting v2.

#### Which issue(s) this PR fixes or relates to

Contrib https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.