mimirpb: Make RW2 metadata serialization order deterministic (#12201)
#### What this PR does

When deserializing RW2 requests, the metadata slice might come back in a
different order for the same request.
This doesn't change the meaning of the in practice, so from the user
perspective the order has no meaning. However, it makes it hard to
compare RW2 requests (e.g. in tests) because serializing the same
request repeatedly is not bit-equal.

This happens because the RW2 metadata is stored into a map by family
name, which is then iterated through to produce the final slice - the
map iteration might yield different orders on different unmarshal
passes, for the same request.

This PR fixes it by remembering the order of the metadata as we parse
the request, then we populate the metadata field in that order.

#### Which issue(s) this PR fixes or relates to

contrib https://github.com/grafana/mimir-squad/issues/2253

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.