MQE: refactor responsibilities of `Engine` and `Query`, introduce `Evaluator` and `Materializer` (#12122)
#### What this PR does

This PR refactors the responsibilities of some of the types in MQE, and
introduces some new types. The goal is to make implementing remote
execution in queriers of query plans created by query-frontends easier.

Main changes:
* The `Materializer` type has been introduced: it owns the
responsibility of converting a query plan into operators (which was
previously the split betwen `Engine` and `Query`)
* The `Evaluator` type has been introduced: it owns the responsibility
of evaluating an operator and passing the data to the evaluation
observer. In the future, it'll support evaluating multiple related
operators from the same query plan.
* The `Query` type now exists primarily to provide an implementation of
the `promql.Query` interface, bridging the interface that Prometheus and
its querying API expect with how MQE operates

There should be no user-visible behaviour changes.

Once we have remote execution of query plans, I expect the interactions
between types to look something like this:
* query-frontends create a query plan by calling
`Engine.NewInstantQuery` or `Engine.NewRangeQuery`, and pass some or all
of the query plan to queriers
* queriers materialize the appropriate part of the query plan with
`Materializer`, pass this to an `Evaluator`, and stream results back to
query-frontends as they are produced by the `Evaluator`

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.